+++ CHANGELOG.md
@@ -16,6 +16,8 @@ sensu-backend init tool.
 - Added the cluster's distribution to Tessen data.
 - Added `--keepalive-critical-timeout` to define the time after which a
 critical keepalive event should be created for an agent.
+- Added `--keepalive-warning-timeout` which is an alias of `--keepalive-timeout`
+for backwards compatibility.
 
 ### Fixed
 - Add a timeout to etcd requests when retrieving the nodes health.